renamed file to utils
Cause thats where the function should live
